requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
decorator==5.1.1
entrypoints==0.4
farm-haystack==1.15.1
filelock==3.10.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
lime==0.2.0.1
mailerlite==0.1.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
openai
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pypdf==3.7.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.2
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
six==1.16.0
spacy==3.7.2
https://huggingface.co/spacy/en_core_web_md/resolve/main/en_core_web_md-any-py3-none-any.whl
smmap==5.0.0
streamlit==1.29.0
streamlit-analytics==0.4.1
streamlit-toggle-switch==1.0.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.3
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
zipp==3.15.0
